requirements/base.txt
# The comment on the next line tells requests.io to warn us if there's a newer
# version of Django within the given range, but not for versions outside that
# range. So if 1.11.12 gets released, we get warned. If 2.0.1 gets released,
# we don't.
Django==2.2.7  # rq.filter: >=2.2.7,<2.3
# Required by Django
sqlparse==0.3.0
pytz
graphene-django==2.6.0
graphene-elastic==0.6
django-filter==2.2.0
dealer==2.0.5
django-dotenv==1.4.2
dj-database-url==0.5.0
Pillow==6.2.0
psycopg2-binary==2.8.4
whitenoise==4.1.4
lxml==4.4.1
libratom==0.2.2
pandas==0.25.3
django-elasticsearch-dsl==7.1.0
django-graphql-jwt==0.3.0
django-simple-history==2.8.0
